backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
pymysql
pillow
oss2
requests
